pyproject.toml
[project]
name = "3d-space-shooter"
version = "0.1.0"
description = "A 3D wireframe space shooter game using PyOpenGL"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pygame>=2.5.0",
    "PyOpenGL>=3.1.7",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
space-shooter = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/game"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"